+++ hata/discord/embed/embed/tests/test__parse_type.py
@@ -4,13 +4,23 @@ from ..fields import parse_type
 from ..preinstanced import EmbedType
 
 
-def test__parse_type():
+def _iter_options():
+    yield {}, EmbedType.rich
+    yield {'type': EmbedType.gifv.value}, EmbedType.gifv
+
+
+@vampytest._(vampytest.call_from(_iter_options()).returning_last())
+def test__parse_type(input_data):
     """
     Tests whether ``parse_type`` works as intended.
+    
+    Parameters
+    ----------
+    input_data : `dict<str, object>`
+        Input data.
+    
+    Returns
+    -------
+    output : ``EmbedType``
     """
-    for input_data, expected_output in (
-        ({}, EmbedType.rich),
-        ({'type': EmbedType.gifv.value}, EmbedType.gifv),
-    ):
-        output = parse_type(input_data)
-        vampytest.assert_is(output, expected_output)
+    return parse_type(input_data)