+++ hata/discord/onboarding/onboarding_screen/fields.py
@@ -2,19 +2,22 @@ __all__ = ()
 
 from ...channel import Channel
 from ...field_parsers import (
-    bool_parser_factory, entity_id_array_parser_factory, entity_id_parser_factory, nullable_object_array_parser_factory
+    bool_parser_factory, entity_id_array_parser_factory, entity_id_parser_factory, nullable_object_array_parser_factory,
+    preinstanced_parser_factory
 )
 from ...field_putters import (
     bool_optional_putter_factory, entity_id_putter_factory, nullable_entity_array_putter_factory,
-    optional_entity_id_array_optional_putter_factory
+    optional_entity_id_array_optional_putter_factory, preinstanced_putter_factory
 )
 from ...field_validators import (
     bool_validator_factory, entity_id_array_validator_factory, entity_id_validator_factory,
-    nullable_object_array_validator_factory
+    nullable_object_array_validator_factory, preinstanced_validator_factory
 )
 from ...guild import Guild
 
 from ..onboarding_prompt import OnboardingPrompt
+
+from .preinstanced import OnboardingMode
 
 # default_channel_ids
 
@@ -33,6 +36,12 @@ validate_enabled = bool_validator_factory('enabled', True)
 parse_guild_id = entity_id_parser_factory('guild_id')
 put_guild_id_into = entity_id_putter_factory('guild_id')
 validate_guild_id = entity_id_validator_factory('guild_id', Guild)
+
+# mode
+
+parse_mode = preinstanced_parser_factory('mode', OnboardingMode, OnboardingMode.default)
+put_mode_into = preinstanced_putter_factory('mode')
+validate_mode = preinstanced_validator_factory('mode', OnboardingMode)
 
 # prompts