+++ hata/discord/onboarding/onboarding_screen/onboarding_screen.py
@@ -2,14 +2,16 @@ __all__ = ('OnboardingScreen',)
 
 from scarletio import RichAttributeErrorBaseType
 
-from .fields import (
-    parse_default_channel_ids, parse_enabled, parse_guild_id, parse_prompts, put_default_channel_ids_into,
-    put_enabled_into, put_guild_id_into, put_prompts_into, validate_default_channel_ids, validate_enabled,
-    validate_guild_id, validate_prompts
-)
-from ...precreate_helpers import process_precreate_parameters_and_raise_extra
-from ...core import GUILDS
 from ...channel import ChannelType, create_partial_channel_from_id
+from ...core import GUILDS
+from ...precreate_helpers import process_precreate_parameters_and_raise_extra
+
+from .fields import (
+    parse_default_channel_ids, parse_enabled, parse_guild_id, parse_mode, parse_prompts, put_default_channel_ids_into,
+    put_enabled_into, put_guild_id_into, put_mode_into, put_prompts_into, validate_default_channel_ids,
+    validate_enabled, validate_guild_id, validate_mode, validate_prompts
+)
+from .preinstanced import OnboardingMode
 
 
 PRECREATE_FIELDS = {
@@ -18,6 +20,7 @@ PRECREATE_FIELDS = {
     'enabled': ('enabled', validate_enabled),
     'guild_id': ('guild_id', validate_guild_id),
     'guild': ('guild_id', validate_guild_id),
+    'mode': ('mode', validate_mode),
     'prompts': ('prompts', validate_prompts),
 }
 
@@ -34,12 +37,14 @@ class OnboardingScreen(RichAttributeErrorBaseType):
         Whether onboarding is enabled.
     guild_id : `int`
         The guild's identifier the onboarding screen is part of.
+    mode : ``OnboardingMode``
+        Onboarding mode.
     prompts : `None`, `tuple` of ``OnboardingPrompt``
         The prompts shown during onboarding and in customize community.
     """
-    __slots__ = ('default_channel_ids', 'enabled', 'guild_id', 'prompts',)
+    __slots__ = ('default_channel_ids', 'enabled', 'guild_id', 'mode', 'prompts')
     
-    def __new__(cls, *, default_channel_ids = ..., enabled = ..., prompts = ...):
+    def __new__(cls, *, default_channel_ids = ..., enabled = ..., mode = ..., prompts = ...):
         """
         Creates an onboarding screen instance from the given parameters.
         
@@ -49,6 +54,8 @@ class OnboardingScreen(RichAttributeErrorBaseType):
             The channels' identifiers that new members get opted into automatically.
         enabled : `bool`, Optional (Keyword only)
             Whether onboarding is enabled.
+        mode : ``OnboardingMode``, `int`, Optional (Keyword only)
+            Onboarding mode.
         prompts : `None`, `iterable` of ``OnboardingPrompt``, Optional (Keyword only)
             The prompts shown during onboarding and in customize community.
         
@@ -71,16 +78,24 @@ class OnboardingScreen(RichAttributeErrorBaseType):
         else:
             enabled = validate_enabled(enabled)
         
+        # mode
+        if mode is ...:
+            mode = OnboardingMode.default
+        else:
+            mode = validate_mode(mode)
+        
         # prompts
         if prompts is ...:
             prompts = None
         else:
             prompts = validate_prompts(prompts)
         
+        # Construct
         self = object.__new__(cls)
         self.default_channel_ids = default_channel_ids
         self.enabled = enabled
         self.guild_id = 0
+        self.mode = mode
         self.prompts = prompts
         return self
     
@@ -103,6 +118,7 @@ class OnboardingScreen(RichAttributeErrorBaseType):
         self.default_channel_ids = parse_default_channel_ids(data)
         self.enabled = parse_enabled(data)
         self.guild_id = parse_guild_id(data)
+        self.mode = parse_mode(data)
         self.prompts = parse_prompts(data)
         return self
     
@@ -125,6 +141,7 @@ class OnboardingScreen(RichAttributeErrorBaseType):
         data = {}
         put_default_channel_ids_into(self.default_channel_ids, data, defaults)
         put_enabled_into(self.enabled, data, defaults)
+        put_mode_into(self.mode, data, defaults)
         put_prompts_into(self.prompts, data, defaults, include_internals = include_internals)
         
         if include_internals:
@@ -169,6 +186,14 @@ class OnboardingScreen(RichAttributeErrorBaseType):
             
             repr_parts.append(']')
         
+        # mode
+        mode = self.mode
+        repr_parts.append(', mode = ')
+        repr_parts.append(mode.name)
+        repr_parts.append(' ~ ')
+        repr_parts.append(str(mode))
+        
+        # prompts
         prompts = self.prompts
         if (prompts is not None):
             repr_parts.append(', prompts = [')
@@ -208,6 +233,9 @@ class OnboardingScreen(RichAttributeErrorBaseType):
         # guild_id
         # ignore, internal
         
+        if self.mode is not other.mode:
+            return False
+        
         if self.prompts != other.prompts:
             return False
         
@@ -218,6 +246,7 @@ class OnboardingScreen(RichAttributeErrorBaseType):
         """Returns the onboarding screen's hash value ."""
         hash_value = 0
         
+        # default_channel_id
         default_channel_ids = self.default_channel_ids
         if (default_channel_ids is not None):
             hash_value ^= len(default_channel_ids)
@@ -225,11 +254,16 @@ class OnboardingScreen(RichAttributeErrorBaseType):
             for channel_id in default_channel_ids:
                 hash_value ^= channel_id
         
+        # enabled
         hash_value ^= self.enabled << 4
         
         # guild_id
         # Ignore it
         
+        # mode
+        hash_value ^= self.mode.value << 11
+        
+        # prompts
         prompts = self.prompts
         if (prompts is not None):
             hash_value ^= len(prompts) << 5
@@ -258,10 +292,12 @@ class OnboardingScreen(RichAttributeErrorBaseType):
             Alternative of `default_channel_ids`.
         enabled : `bool`, Optional (Keyword only)
             Whether onboarding is enabled.
+        guild : `int`, ``Guild``, Optional (Keyword only)
+            Alternative of `guild_id`.
         guild_id : `int`, ``Guild``, Optional (Keyword only)
             The guild's identifier the onboarding screen is part of.
-        guild_id : `int`, ``Guild``, Optional (Keyword only)
-            Alternative of `guild_id`.
+        mode : ``OnboardingMode``, `int`, Optional (Keyword only)
+            Onboarding mode.
         prompts : `None`, `iterable` of ``OnboardingPrompt``, Optional (Keyword only)
             The prompts shown during onboarding and in customize community.
         
@@ -280,6 +316,7 @@ class OnboardingScreen(RichAttributeErrorBaseType):
         self.default_channel_ids = None
         self.enabled = True
         self.guild_id = 0
+        self.mode = OnboardingMode.default
         self.prompts = None
         
         if (processed is not None):
@@ -306,6 +343,7 @@ class OnboardingScreen(RichAttributeErrorBaseType):
         
         new.enabled = self.enabled
         new.guild_id = 0
+        new.mode = self.mode
         
         prompts = self.prompts
         if (prompts is not None):
@@ -315,18 +353,18 @@ class OnboardingScreen(RichAttributeErrorBaseType):
         return new
     
     
-    def copy_with(self, *, default_channel_ids = ..., enabled = ..., prompts = ...):
+    def copy_with(self, *, default_channel_ids = ..., enabled = ..., mode = ..., prompts = ...):
         """
         Copies the onboarding screen with the given screen.
         
-        Parameters
-        ----------
         Parameters
         ----------
         default_channel_ids : `None`, `iterable` of (`int`, ``Channel``), Optional (Keyword only)
             The channels' identifiers that new members get opted into automatically.
         enabled : `bool`, Optional (Keyword only)
             Whether onboarding is enabled.
+        mode : ``OnboardingMode``, `int`, Optional (Keyword only)
+            Onboarding mode.
         prompts : `None`, `iterable` of ``OnboardingPrompt``, Optional (Keyword only)
             The prompts shown during onboarding and in customize community.
         
@@ -355,6 +393,12 @@ class OnboardingScreen(RichAttributeErrorBaseType):
         else:
             enabled = validate_enabled(enabled)
         
+        # mode
+        if mode is ...:
+            mode = self.mode
+        else:
+            mode = validate_mode(mode)
+        
         # prompts
         if prompts is ...:
             prompts = self.prompts
@@ -367,6 +411,7 @@ class OnboardingScreen(RichAttributeErrorBaseType):
         new.default_channel_ids = default_channel_ids
         new.enabled = enabled
         new.guild_id = 0
+        new.mode = mode
         new.prompts = prompts
         return new