+++ hata/discord/onboarding/onboarding_screen/tests/test__OnboardingScreen__data.py
@@ -3,6 +3,7 @@ import vampytest
 from ...onboarding_prompt import OnboardingPrompt
 
 from ..onboarding_screen import OnboardingScreen
+from ..preinstanced import OnboardingMode
 
 from .test__OnboardingScreen__constructor import _assert_fields_set
 
@@ -15,6 +16,7 @@ def test__OnboardingScreen__from_data__0():
     """
     default_channel_ids = [202303040036, 202303040037]
     enabled = True
+    mode = OnboardingMode.advanced
     prompts = [
         OnboardingPrompt.precreate(202303040038, name = 'ibuki'),
         OnboardingPrompt.precreate(202303040039, name = 'suika'),
@@ -25,6 +27,7 @@ def test__OnboardingScreen__from_data__0():
         'default_channel_ids': [str(channel_id) for channel_id in default_channel_ids],
         'enabled': enabled,
         'guild_id': str(guild_id),
+        'mode': mode.value,
         'prompts': [prompt.to_data(defaults = True, include_internals = True) for prompt in prompts],
     }
     
@@ -33,6 +36,7 @@ def test__OnboardingScreen__from_data__0():
     
     vampytest.assert_eq(screen.default_channel_ids, tuple(default_channel_ids))
     vampytest.assert_eq(screen.enabled, enabled)
+    vampytest.assert_is(screen.mode, mode)
     vampytest.assert_eq(screen.prompts, tuple(prompts))
     vampytest.assert_eq(screen.guild_id, guild_id)
 
@@ -46,6 +50,7 @@ def test__OnboardingScreen__to_data__0():
     """
     default_channel_ids = [202303040041, 202303040042]
     enabled = True
+    mode = OnboardingMode.advanced
     prompts = [
         OnboardingPrompt.precreate(202303040043, name = 'ibuki'),
         OnboardingPrompt.precreate(202303040044, name = 'suika'),
@@ -56,6 +61,7 @@ def test__OnboardingScreen__to_data__0():
         default_channel_ids = default_channel_ids,
         enabled = enabled,
         prompts = prompts,
+        mode = mode,
         guild_id = guild_id,
     )
     
@@ -63,6 +69,7 @@ def test__OnboardingScreen__to_data__0():
         'default_channel_ids': [str(channel_id) for channel_id in default_channel_ids],
         'enabled': enabled,
         'guild_id': str(guild_id),
+        'mode': mode.value,
         'prompts': [prompt.to_data(defaults = True, include_internals = True) for prompt in prompts],
     }